section11/aws/wfdb/requirements.txt
amqp==2.1.1
aniso8601==1.1.0
appnope==0.1.0
awsebcli==3.8.4
backports.shutil-get-terminal-size==1.0.0
billiard==3.5.0.2
blessed==1.9.5
blinker==1.4
botocore==1.4.78
celery==4.0.0
cement==2.8.2
click==6.6
colorama==0.3.7
coverage==4.2
cssmin==0.2.0
decorator==4.0.10
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.12
Flask==0.11.1
Flask-Admin==1.4.2
Flask-Assets==0.12
Flask-Caching==1.0.1
Flask-Celery-Helper==1.1.0
Flask-DebugToolbar==0.10.0
Flask-Login==0.4.0
Flask-Principal==0.4.0
Flask-RESTful==0.3.5
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-WTF==0.13.1
ipython==5.1.0
ipython-genutils==0.1.0
itsdangerous==0.24
Jinja2==2.8
jmespath==0.9.0
kombu==4.0.0
MarkupSafe==0.23
pathspec==0.3.4
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.9
psycopg2==2.6.2
ptyprocess==0.5.1
Pygments==2.1.3
python-dateutil==2.5.3
pytz==2016.7
PyYAML==3.12
redis==2.10.5
requests==2.9.1
semantic-version==2.5.0
simplegeneric==0.8.1
six==1.10.0
SQLAlchemy==1.1.4
texttable==0.8.7
traitlets==4.3.1
vine==1.1.3
wcwidth==0.1.7
webassets==0.12.0
websocket-client==0.37.0
Werkzeug==0.11.11
WTForms==2.1
